+++ python/triton_kernels/triton_kernels/routing.py
@@ -5,6 +5,8 @@ from .routing_details._routing_compute import _routing_memset_indx
 from .routing_details._routing_compute import _routing_compute_indx_offs
 from .routing_details._routing_compute import _routing_compute_indx
 from .routing_details._routing_compute import _routing_clear_bitmatrix
+from .routing_details._expt_data import _expt_data_memset
+from .routing_details._expt_data import _expt_data_compute
 
 
 @dataclass
@@ -30,11 +32,34 @@ class ScatterIndx:
 
 
 @dataclass
+class ExptData:
+    # hist[i] is the number of tokens routed to expert i
+    hist: torch.Tensor
+    # token_offs_raw[i] is the offset of the first token routed
+    # to expert i in an expert-sorted array
+    token_offs_raw: torch.Tensor
+    # token_offs_pad[block][i] is the offset of the first token routed
+    # to expert i in an expert-sorted array, assuming histogram
+    # rounded to the next multiple of `block`
+    token_offs_pad: dict[int, torch.Tensor]
+    # block_id_map[block] contain one value for each `pid`` launched by
+    # the matrix multiplication kernel launched with BLOCK_M=block:
+    # - the value is -1 if the `pid` has no work to do
+    # - otherwise, the value is two int16 (packed as an int32) that
+    #   correspond respectively to (1) the expert assigned to
+    #   the tokens processed by this pid; (2) the block assigned to the
+    #   tokens processed by this pid (think `pid_m` in a regular matmul)
+    # see `test_routing.py` for a reference implementation and more details
+    block_pid_map: dict[int, torch.Tensor]
+
+
+@dataclass
 class RoutingData:
     gate_scal: torch.Tensor = field()
     expt_hist: torch.Tensor = field()
     n_expts_tot: int = field()
     n_expts_act: int = field()
+    expt_data: ExptData = None
 
     # Used to make perf annotation cleaner: when we use expert sharding, we can
     # use this to tell the "expected" number of local tokens per expert, because
@@ -145,6 +170,58 @@ def prune_routing(expt_scal, expt_indx, bitmatrix, simulated_ep):
 
 
 # --------------------------
+# expt_data
+# --------------------------
+
+
+def log2_power_of_two(x):
+    assert x > 0 and (x & (x - 1)) == 0, "x must be a power of two"
+    return x.bit_length() - 1
+
+
+def compute_expt_data(expt_hist, n_expts_tot, n_gates):
+    if expt_hist is None:
+        return ExptData(None, None, None, None)
+    MEMSET_BLOCK = 128
+    HIST2_BLOCK_M = 512
+    device = expt_hist.device
+    n_expts_tot = n_expts_tot
+    cdiv = triton.cdiv
+    # block_ms are all powers-of-two between 16 and 128 (inclusive)
+    block_m_log2_start = 4
+    block_m_log2_end = 8
+    block_m_num = block_m_log2_end - block_m_log2_start
+    if n_gates <= n_expts_tot:
+        max_n_tiles = n_gates
+    else:
+        max_n_tiles = n_expts_tot - 1 - ((n_expts_tot - n_gates - 1) // 2**block_m_log2_start)
+    # allocate memory
+    pad = lambda x: cdiv(x, MEMSET_BLOCK) * MEMSET_BLOCK
+    dtype = torch.int32
+    token_offs_raw = torch.empty((n_expts_tot + 1, ), dtype=dtype, device=device)
+    token_offs_pad = torch.empty((block_m_num, pad(n_expts_tot + 1)), dtype=dtype, device=device)
+    block_pid_map = torch.empty((block_m_num, pad(max_n_tiles)), dtype=dtype, device=device)
+    # compute outputs
+    token_offs_pad = token_offs_pad[:, :n_expts_tot + 1]
+    block_pid_map = block_pid_map[:, :max_n_tiles]
+    memset_grid = cdiv(block_pid_map.shape[1], MEMSET_BLOCK) + 1
+    _expt_data_memset[(memset_grid, block_m_num)](
+        expt_hist, n_expts_tot, token_offs_raw,  #
+        token_offs_pad, token_offs_pad.stride(0),  #
+        block_pid_map, block_pid_map.stride(0),  #
+        block_m_log2_start, BLOCK=MEMSET_BLOCK,  # optimization parameters
+        num_warps=1)
+    _expt_data_compute[(n_expts_tot, block_m_num)](
+        expt_hist, token_offs_pad, token_offs_pad.stride(0), block_pid_map, block_pid_map.stride(0),  # outputs
+        block_m_log2_start, BLOCK=HIST2_BLOCK_M,  # optimization parameters
+        num_warps=4)
+    # unpack into datastructure
+    token_offs_pad = {2**j: token_offs_pad[i, :] for i, j in enumerate(range(block_m_log2_start, block_m_log2_end))}
+    block_pid_map = {2**j: block_pid_map[i, :] for i, j in enumerate(range(block_m_log2_start, block_m_log2_end))}
+    return ExptData(expt_hist, token_offs_raw, token_offs_pad, block_pid_map)
+
+
+# --------------------------
 # routing
 # --------------------------
 
@@ -163,10 +240,48 @@ def routing(logits, n_expts_act, sm_first=False, expt_indx=None, simulated_ep=1,
     n_expts_tot = logits.shape[-1] // simulated_ep
     gather_indx = GatherIndx(src_indx=topk_indx, dst_indx=gate_indx)
     scatter_indx = ScatterIndx(src_indx=gate_indx, dst_indx=topk_indx)
-    return RoutingData(gate_scal, hist, n_expts_tot, n_expts_act), gather_indx, scatter_indx
+    print("triton: ", logits.shape[0], n_expts_act)
+    expt_data = compute_expt_data(hist, n_expts_tot, topk_indx.numel())
+    return RoutingData(gate_scal, hist, n_expts_tot, n_expts_act, expt_data), gather_indx, scatter_indx
+
+
+# --------------------------
+# torch reference
+# --------------------------
+
+
+def compute_expt_data_torch(hist, n_expts_tot, n_gates):
+    # offset for each experts
+    device = hist.device
+    token_offs_raw = torch.cumsum(hist, dim=0)
+    token_offs_raw = torch.cat((torch.zeros(1, dtype=torch.int32, device=device), token_offs_raw))
+    # maximum number of tiles for all values of `block_m` considered
+    block_ms = [16, 32, 64, 128]
+    if n_gates <= n_expts_tot:
+        max_n_tiles = n_gates
+    else:
+        # ceil_div(n_gates - n_experts + 1, d_tile) + n_experts - 1
+        # ceil_div(x, y): -(-x // y)
+        max_n_tiles = n_expts_tot - 1 - ((n_expts_tot - n_gates - 1) // min(block_ms))
+    # fill up tile offset/infos for each block
+    token_offs_pad = dict()
+    block_pid_map = dict()
+    for block_m in [16, 32, 64, 128]:
+        n_tiles = (hist + block_m - 1) // block_m  # matmul blocks needed
+        token_offs_pad[block_m] = torch.cumsum(n_tiles, dim=0)
+        token_offs_pad[block_m] = torch.cat((torch.zeros(1, dtype=torch.int32, device=device), token_offs_pad[block_m]))
+        # compute data required to drive ragged batch matmul
+        block_pid_map[block_m] = -torch.ones(max_n_tiles, dtype=torch.int32, device=device)
+        for e in range(n_expts_tot):
+            offset = token_offs_pad[block_m][e]
+            for b in range(n_tiles[e]):
+                block_pid_map[block_m][offset + b] = (b << 16) + e
+    return ExptData(hist, token_offs_raw, token_offs_pad, block_pid_map)
 
 
 def routing_torch(logits, n_expts_act, sm_first=False, expt_indx=None, n_rows=None):
+    n_gates_pad = logits.shape[0] * n_expts_act
+
     if n_rows is not None:
         logits = logits[:n_rows, :]
 
@@ -196,4 +311,6 @@ def routing_torch(logits, n_expts_act, sm_first=False, expt_indx=None, n_rows=No
     # pack the matmul data structure
     gather_indx = GatherIndx(src_indx=topk_indx.int(), dst_indx=gate_indx.int())
     scatter_indx = ScatterIndx(src_indx=gate_indx.int(), dst_indx=topk_indx.int())
-    return RoutingData(gate_scal, hist, n_expts_tot, n_expts_act), gather_indx, scatter_indx
+    # compute expt_data
+    expt_data = compute_expt_data_torch(hist, n_expts_tot, n_gates_pad)
+    return RoutingData(gate_scal, hist, n_expts_tot, n_expts_act, expt_data), gather_indx, scatter_indx
